+++ test/vhpi/vhpi_test.h
@@ -60,6 +60,7 @@ void vhpi17_startup(void);
 void vhpi18_startup(void);
 void vhpi19_startup(void);
 void vhpi21_startup(void);
+void vhpi22_startup(void);
 void issue744_startup(void);
 void issue762_startup(void);
 void issue978_startup(void);